third_party/WebKit/LayoutTests/fast/dom/documentmarker-set-active-expected.txt

CONSOLE ERROR: line 19: Uncaught SyntaxError: Illegal return statement
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
text
